goreleaser
changed repo structure to support goreleaser
