+++ ca_config/templates/csr.json
@@ -11,6 +11,6 @@
       "OU": "Container lab"
     }],
     "hosts": [
-      "{{.Name}}.{{.Prefix}}.io"
+      "{{.Name}}"
     ]
   }
